fix(db): longin/longout ADEL/MDEL are DBF_LONG — reject over-range puts, don't saturate
Cause 1 of the range-put family. longin/longout model their ADEL/MDEL
deadbands as f64, but the .dbd declares `field(ADEL,DBF_LONG)` /
`field(MDEL,DBF_LONG)`. The client-put target resolves from the STORED type
(get_field's variant), so the f64 storage keyed the string->native parse on
DBF_DOUBLE: c_parse::put_string ran the float row, ACCEPTED `2147483648`
(2^31, past i32::MAX), and the read path saturated it to i32::MAX — where C's
putStringLong -> epicsParseInt32 REFUSES the put and the field keeps its old
value. 8 oracle put-accept defects (longin/longout . ADEL/MDEL . over-max/
under-min).

Structural fix, mirroring int64in/int64out Cause B (commit 224d5ad5): store
ADEL/MDEL in their C-declared i32 variant, so get_field reports Long, the put
target becomes DBF_LONG, and the existing single owner c_parse::put_string
parses via the integer row — refusing over-i32/under-i32, keeping strtol
trailing-text tolerance. No range logic reimplemented. The deadband compare
(check_deadband_ext) reads the field via to_f64 unchanged; MLST/ALST stay f64
(they hold the last VAL, a wider quantity, exactly as before).

Regression: longin_longout_deadband_is_dbf_long.rs — over-i32 refused, band
edge (i32::MAX/MIN) and trailing text accepted.

+++ crates/epics-base-rs/src/server/records/longin.rs
@@ -28,11 +28,18 @@ pub struct LonginRecord {
     // `self.record.get_field("LALM")` — so it stays.
     #[field(type = "Double")]
     pub lalm: f64,
-    // Deadband
-    #[field(type = "Double")]
-    pub adel: f64,
-    #[field(type = "Double")]
-    pub mdel: f64,
+    // Deadband. ADEL/MDEL are `DBF_LONG` (longinRecord.dbd.pod) — the archive
+    // and monitor deadbands are integer counts, not doubles. Stored as `i32`
+    // (not `f64`) so the client-put target resolves to `DBF_LONG` and a string
+    // put parses through the single `c_parse::put_string` owner's `Long` row,
+    // which REFUSES an over-i32/under-i32 value exactly as C's `epicsParseInt32`
+    // does — instead of the `DBF_DOUBLE` row accepting it and the read-path
+    // projection saturating it to `i32::MAX`/`MIN`. Mirrors Cause A/B of the
+    // put-conversion family (int64in/int64out DBF_INT64, commit 224d5ad5).
+    #[field(type = "Long")]
+    pub adel: i32,
+    #[field(type = "Long")]
+    pub mdel: i32,
     // Alarm-range time-constant filter (longinRecord.c::checkAlarms:310-356).
     // AFTC > 0 low-pass-filters the integer alarmRange so transient
     // excursions don't immediately alarm; AFVL is the accumulator.
@@ -78,8 +85,8 @@ impl Default for LonginRecord {
             hopr: 0,
             lopr: 0,
             lalm: 0.0,
-            adel: 0.0,
-            mdel: 0.0,
+            adel: 0,
+            mdel: 0,
             aftc: 0.0,
             afvl: 0.0,
             alst: 0.0,
